Reorganise test infrastructure
